backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pgvector
openai
python-multipart
pytest
ruff
mypy
nox
pre-commit
torch
torchvision
pillow
scikit-learn
jsonschema
types-jsonschema
PyYAML
